+++ packages/wallet-frontend/src/components/kyc/PersonalDetailsForm.tsx
@@ -10,20 +10,21 @@ import { fetchCountries, fetchDocuments, getObjectKeys } from '@/utils/helpers'
 import { useEffect, useState } from 'react'
 import { SuccessDialog } from '../dialogs/SuccessDialog'
 import { useKYCFormContext } from './context'
+import { Controller } from 'react-hook-form'
 
 export const PersonalDetailsForm = () => {
   const [openDialog, closeDialog] = useDialog()
   const [countries, setCountries] = useState<SelectOption[]>([])
   const { setTab, setDisabled, setIdTypes } = useKYCFormContext()
 
-  // set default values testing, Denmark is selected for ID verification and because it supports multiple currencies
-  const defaultCountry = USE_TEST_DATA_KYC
-    ? { value: 'DK', name: 'Denmark' }
-    : undefined
   const defaultTestValues = {
     city: 'Copenhagen',
     address: 'Den Lille Havfrue',
-    zip: '2100'
+    zip: '2100',
+    country: {
+      value: 'DK',
+      label: 'Denmark'
+    }
   }
 
   const personalDetailsForm = useZodForm({
@@ -43,6 +44,7 @@ export const PersonalDetailsForm = () => {
     getCountries()
   }, [])
 
+console.log(personalDetailsForm.formState.errors)
   return (
     <Form
       form={personalDetailsForm}
@@ -95,16 +97,27 @@ export const PersonalDetailsForm = () => {
           label="Last name"
         />
       </div>
-      <Select
+      <Controller
         name="country"
-        setValue={personalDetailsForm.setValue}
-        error={personalDetailsForm.formState.errors.country?.message}
-        options={countries}
-        label="Country"
-        // TESTING => always disabled and Denmark selected
-        defaultValue={defaultCountry}
-        isDisabled={USE_TEST_DATA_KYC}
+        render={({ field: { value } }) => (
+          <Select<SelectOption>
+            options={countries}
+            value={value}
+            onChange={(option) => {
+              if (option) {
+                personalDetailsForm.setValue(
+                  'country',
+                  { ...option },
+                  {
+                    shouldDirty: true
+                  }
+                )
+              }
+            }}
+          />
+        )}
       />
+      {personalDetailsForm.formState.errors.country?.message}
       <Input
         required
         {...personalDetailsForm.register('city')}
